# Conflicts:
#   frontend/src/components/Community/CommunityDetail.vue

+++ backend/src/main/java/com/mococo/common/service/PostService.java
@@ -199,17 +199,31 @@ public class PostService {
 	}
 
 	public Post updatePost(Post post, MultipartFile[] files, List<Integer> dlist) throws IllegalStateException, IOException{
-		
+		Optional<PostPhoto> postphoto =null;
 		// 삭제할 이미지 리스트들 삭제.
-		for(Integer photono : dlist) {
-			postphotoDAO.deleteById(photono);
+		if(dlist == null) {
+			
 		}
-		
+		else {
+			for(Integer photono : dlist) {
+				//먼저 번호로 Image의 고유 이름을 찾는다.
+				postphoto = postphotoDAO.findById(photono);
+				
+				// Bucket에서 삭제한다.
+				amazonS3.deleteObject(s3bucket, postphoto.get().getSaveFolder()+"/"+postphoto.get().getSaveFile());
+				
+				// db에서도 정보를 삭제한다.
+				postphotoDAO.deleteById(photono);
+			}
+			
+		}
+
 		
 		Post p = postDAO.save(post);
 		
         if(files == null){
             // TODO : 파일이 없을 땐 어떻게 해야할까.. 고민을 해보아야 할 것
+        	System.out.println("텅비었어....");
         }
         // 파일에 대해 DB에 저장하고 가지고 있을 것
         else{
@@ -219,25 +233,22 @@ public class PostService {
 				if (!originalFileName.isEmpty()) {
 					String sourceFileName = mfile.getOriginalFilename();
 					String sourceFileNameExtension = FilenameUtils.getExtension(sourceFileName).toLowerCase();
-					File destinationFile;
+
 					String destinationFileName;
-					String fileUrl = "C:\\SSAFY\\Mococo\\backend\\src\\main\\resources\\photos\\";
-					do {
-						destinationFileName = RandomStringUtils.randomAlphanumeric(32) + "." + sourceFileNameExtension;
-						destinationFile = new File(fileUrl + destinationFileName);
-					} while (destinationFile.exists());
+					destinationFileName = RandomStringUtils.randomAlphanumeric(32) + "." + sourceFileNameExtension;
 	
-					destinationFile.getParentFile().mkdirs();
-					mfile.transferTo(destinationFile);
 	
-					photo.setSaveFile(destinationFileName);
-					photo.setOriginFile(sourceFileName);
-					photo.setSaveFolder(fileUrl);
+					// S3 Bucket에 저장
+					File file = convertMultiPartFileToFile(mfile);
 					
-					System.out.println("길이" + photo.getSaveFolder().length());
+					amazonS3.putObject(new PutObjectRequest(s3bucket, "post/"+destinationFileName, file).withCannedAcl(CannedAccessControlList.PublicRead));
+					
 					photo.setPost(post);
-					System.out.println(photo.getSaveFile());
+					photo.setOriginFile(originalFileName);
+					photo.setSaveFile(destinationFileName);
+					photo.setSaveFolder("post");
 					postphotoDAO.save(photo);
+					file.delete();
 				}
 
 			}
@@ -263,6 +274,15 @@ public class PostService {
         	System.out.println("Error converting the multi-part file to file= "+ex.getMessage());
         }
         return file;
+	}
+
+	public List<Object> findLikePostById(int user_number) {
+		List<Object> ret = postrecommendDAO.findLikePostById(user_number);
+		
+		
+		
+		
+		return ret;
 	}
